==> 0_extract_data.ipynb <==
import pandas as pd
import numpy as np
import dask.dataframe as dd
import sys
import datetime
import seaborn as sns
import matplotlib.pyplot as plt

# code for extracting data


filename = "thegurus-opendata-renfe-trips.csv"

price_changed = pd.DataFrame([])
count = 0
for df in pd.read_csv(filename, iterator=True, chunksize=10 ** 6):
    count += 1

    df = df.loc[
        (df["meta"] != "{}")
        & (df["vehicle_type"] == "AVE")
        & (df["departure"] > '2020-03-02 07:30:00')
        & (df["departure"] < '2020-03-31 18:45:01')
    ]

    price_changed = pd.concat([price_changed, df])
    print(
        "From {c} million rows, ".format(c=count)
        + str(price_changed.shape[0])
        + " extracted"
    )
    if price_changed.shape[0] > 3 * (10 ** 6):
        break

price_changed.shape

price_changed.to_parquet("afterextract_data.parquet")
