==> 1_data_cleaning.ipynb <==
import pandas as pd
import numpy as np
import datetime
from tqdm import tqdm

data = pd.read_parquet("afterextract_data.parquet").reset_index(drop = True)

data.head()

# generate train_id
data["train_id"] = data[["origin", "destination", "departure","arrival"]].sum(axis=1)

data["weekday"] = data["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").weekday()
)
data["depart_month"] = data["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").month
)
data["depart_hour"] = data["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").hour
)
# only jan, feb, march,
data = data.loc[data["depart_month"] < 4].reset_index(drop=True)

data["depart_year"] = data["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").year
)

data["depart_year"].value_counts()

len(data["train_id"].unique())

# átvett függvény a meta, változóból való információ kinyerésére,

def extract_meta(train):
    prices_seats = pd.DataFrame()
    for i in tqdm(range(train.shape[0])):
        dic_in = json.loads(train.meta[i])
        # extract values from tree dictionary
        while type(list(dic_in.values())[0]) != float:
            res = {key: list(value) for key, value in dic_in.items()}
            dic_out = {}
            for key in res.keys():
                for value in res[key]:
                    dic_out[key + "_" + value] = dic_in[key][value]
            dic_in = dic_out

        # check for seats, if no seat its the price
        for key in list(dic_in.keys()):
            if "seats" not in key:
                if "price" not in key:
                    dic_in[key + "_price"] = dic_in[key]
                    dic_in.pop(key)

        prices_seats = pd.concat(
            [prices_seats, pd.DataFrame.from_dict(dic_in, orient="index").T]
        ).reset_index(drop=True)
    return prices_seats

(data.shape[0]//10000)*10000

#taking tons of time.
chunks = {
    i: pd.concat(
        [
            data.loc[i * 10000 : i * 10000 + 9999, :].reset_index(drop=True),
            extract_meta(
                data.loc[i * 10000 : i * 10000 + 9999, :].reset_index(drop=True)
            ),
        ],
        axis=1,
    )
    for i in range(data.shape[0]//10000)
}

df = pd.concat(list(chunks.values()))

last_chunk = pd.concat(
    [
        data.loc[(data.shape[0]//10000)*10000:, :].reset_index(drop=True),
        extract_meta(data.loc[(data.shape[0]//10000)*10000:, :].reset_index(drop=True)),
    ]
)

df = pd.concat([df,last_chunk])

df.shape

df = df.dropna(subset=["id"]).reset_index(drop=True)

df.shape

df["days_till_dep"] = [
    (
        datetime.datetime.strptime(df.loc[i, "departure"][0:19], "%Y-%m-%d %H:%M:%S")
        - datetime.datetime.strptime(
            df.loc[i, "insert_date"][0:19], "%Y-%m-%d %H:%M:%S"
        )
    ).days
    for i in range(df.shape[0])
]

df["timedelta_till_dep"] = [
    (
        datetime.datetime.strptime(df.loc[i, "departure"][0:19], "%Y-%m-%d %H:%M:%S")
        - datetime.datetime.strptime(
            df.loc[i, "insert_date"][0:19], "%Y-%m-%d %H:%M:%S"
        )
    )
    for i in range(df.shape[0])
]

df.loc[:,["timedelta_till_dep","days_till_dep"]]

df.to_pickle("all_trains.pickle")

___
