==> 2_data_cleaning.ipynb <==
import pandas as pd
import numpy as np
import dask.dataframe as dd
import sys
import datetime
import seaborn as sns
import matplotlib.pyplot as plt

trains = pd.read_pickle("all_trains.pickle")

trains["train_id"] = trains[["origin", "destination", "departure","arrival"]].sum(axis=1)

columns_to_select = [
    "train_id",
    "departure",
    "arrival",
    "duration",
    "vehicle_class",
    "Turista_Promo +_seats",
    "Turista_Promo +_price",
    "insert_date",
    "timedelta_till_dep",
    "origin",
    "destination",
]
df = trains[columns_to_select]

trains["vehicle_class"].value_counts()

df["vehicle_class"].unique()

df = df[df["vehicle_class"] == "Turista"]

df["weekday"] = df["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").weekday()
)
df["depart_month"] = df["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").month
)
df["depart_hour"] = df["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").hour
)

df["depart_day"] = df["departure"].apply(
    lambda x: datetime.datetime.strptime(x, "%Y-%m-%d %H:%M:%S").day
)

def stripstring(x):
    x = x[:10]
    return x

df["insert_date_2"] = df["insert_date"].apply(lambda x: stripstring(x))

df["insert_date_month"] = df["insert_date_2"].apply(
    lambda x: str(datetime.datetime.strptime(x, "%Y-%m-%d").month)
)
df["insert_date_day"] = df["insert_date_2"].apply(
    lambda x: str(datetime.datetime.strptime(x, "%Y-%m-%d").day)
)

def insert_month_day(month,day):
    if len(day) == 1:
        day = "0" + day
    return month + day

def add_0(day):
    if len(day) == 1:
        day = "0" + day
    return day

insert_month_day("2", "10")

df["insert_date_day"] = df["insert_date_day"].apply(add_0)


df["insert_date_day"].unique()

df["insert_month_day"] = df["insert_date_month"] + df["insert_date_day"]
df["insert_month_day"].unique()

df.columns

data = (
    df.groupby(["train_id","insert_month_day"])
    .agg(
        mean_econ_price=("Turista_Promo +_price", np.nanmean),
        max_econ_seats=("Turista_Promo +_seats", max),
        min_econ_seats=("Turista_Promo +_seats", min),
        mean_econ_seats=("Turista_Promo +_seats", "mean"),
        depart_month=("depart_month","first"),
        depart_day=("depart_day", "first"),
        depart_hour=("depart_hour", "first"),
        insert_date = ("insert_date_2", "first"),
        departure = ("departure","first"),
        origin = ("origin","first"),
        destination = ("destination","first"),
    )
    .reset_index()
)

data.shape

data["train_id"].nunique()

data = data[data['mean_econ_seats'].notna()]

data["train_id"].nunique()

data["insert_month_day"].unique()

data.head(2)

data["price_change"] = (
    data.sort_values(by=["train_id", "insert_date"], ascending=True)
    .groupby("train_id")["mean_econ_price"]
    .transform("diff")
)

data["price_change_percentage"] = (data["price_change"] / data["mean_econ_price"]).round(3)

data["price_change_direction"] = np.where(
    data["price_change_percentage"] > 0,
    1,
    np.where(data["price_change_percentage"] < 0, -1, 0),
)

data["price_increase"] = np.where(data["price_change_direction"] == 1, True, False)
data["price_decrease"] = np.where(data["price_change_direction"] == -1, True, False)

data["seats_sold"] = abs((
    data.sort_values(by=["train_id", "insert_date"], ascending=True)
    .groupby("train_id")["mean_econ_seats"]
    .transform("diff")
))

data["price_increase"].value_counts()

data["depart_monthday"] = data["departure"].apply(lambda date: date[5:10])

dfs = [
    x
    for _, x in data[data["insert_month_day"] != "206"].groupby(
        ["insert_month_day", "depart_monthday", "origin", "destination"]
    )
]

from tqdm.notebook import tqdm

for i in tqdm(dfs):
    number_lags = 1
    for lag in range(1, number_lags + 1):
        i["seats_sold_lag" + str(lag)] = i.seats_sold.shift(lag)
        i["pair_" + str(lag)] = i.train_id.shift(lag)

    number_lags_back = 1
    for lag in range(1, number_lags_back + 1):
        i["seats_sold_lag_backward" + str(lag)] = i.seats_sold.shift(-lag)
        i["pair_2"] = i.train_id.shift(-lag)

    i["prop_values"] = i["seats_sold"] / (
        i["seats_sold"]
        + i["seats_sold_lag1"].replace(np.nan, 0)
        + i["seats_sold_lag_backward1"].replace(np.nan, 0)
    )

teljes_data = pd.concat(dfs)

teljes_data.shape

teljes_data.to_csv("data_with_prop_ext.csv")
